% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRndFoReg.R
\name{mlfun}
\alias{mlfun}
\title{MLRndFoReg}
\usage{
mlfun(X, y)
}
\arguments{
\item{X}{A vector of vectors which contain values of independent variables.}

\item{y}{A vector of dependent variables corresponding to the dependent.}
}
\value{
The model with all of its parameters
}
\description{
Takes in dependent variables y and independent variables X and creates a predictive model and gives the stats for the predictions.
}
