% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotfun.R
\name{onehotfun}
\alias{onehotfun}
\title{onehotfun}
\usage{
onehotfun(seqs)
}
\arguments{
\item{seqs}{sequences of 20 nt in length.}
}
\value{
A dataframe with 20 columns where each column represents
a position in the sequence.
}
\description{
Takes in nucleotide sequences of 20 nt in lenght and returns a dataframe
where each row a one hot encoded version of that sequence with each
nucleotide on a separate column
}
\examples{
seqs <- c("ATGCGTACGTAGCTAGCTAG", "CGTACGTAGCTAGCTAGCTA")
onehotdf <- onehotfun(seqs)
head(onehotdf)

}
