% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PkgAvaProj-package.R
\docType{package}
\name{PkgAvaProj-package}
\alias{PkgAvaProj}
\alias{PkgAvaProj-package}
\title{PkgAvaProj: Tools for Advanced Bioinformatics Course}
\description{
This package provides tools and instructions for data preprocessing and analysis in the context of the course Advanced Bioinformatics at Umeå University. The goal of the project is to identify optimal spacer sequences for gene knockout using CRISPR-Cas9.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Sladderdask.github.io/PkgAvaProj}
}

}
\author{
\strong{Maintainer}: Dennis Harding \email{dennis.harding2@gmail.com}

Authors:
\itemize{
  \item Moa Ögren \email{moaa.ogren@gmail.com}
}

}
\keyword{internal}
